Orfox: hook up default panic trigger to "quit and clear"
Signed-off-by: Amogh Pradeep <amoghbl1@gmail.com>

Also:
Bug 28507: Implement fallback to delete private data in the browser startup

When the TBA is forcefully closed, its private data is not deleted,
even if the history.clear_on_exit is set.

As fallback, this patch calls the Sanitize:ClearData event in the
browser startup to clean the private data if needed.