Bug 1450449 - Part 1: Add FileProvider. r=jchen
In case we change our thinking on launching of downloaded files and start using
content:// URIs for that case as well, we already allow our FileProvider to
generate URIs for the whole file system using <root-path>. This is because users
can in principle move our download directory to an arbitrary location on the
file system as long as it is accessible to Firefox. However not all of these
locations (e.g. on a removable SD card) can be specified through the other
methods of specifying available files for a FileProvider, so only the <root-
path> option remains.

MozReview-Commit-ID: 2UStBlU4JsG

--HG--
extra : rebase_source : 2c1828e063c1b3e772ac20c415fd34d0da1c24a6