Bug 28051 - Integrate Orbot and add dependencies
Also:
Bug 28051 - Launch Orbot if it isn't running in the background

Bug 28329 - Part 2. Implement checking if the Tor service is running

Bug 28329 - Part 3. Remove OrbotActivity dependency

Bug 28051 - Stop the background service when we're quitting

If the user swips away the app, then initiate quitting as if the user
selected Quit from the menu.