Bug 27503: Disabling accessibility on Windows breaks screen readers
This patch is merely an application of the current accessibility fixes
in Mozilla 1520177 (
https://bugzilla.mozilla.org/show_bug.cgi?id=1520177 ). It replaces the
RT_MANIFEST symbol (defined in windows headers we don't include) with
the integer constant 24 to avoid includes confusion, updates the
IA2Marshall.rc file to explicitly include the dll manifest and it adds
some additional include directories previously excluded when building
with the mingw toolchain.

With this patch applied, the NVDA screen reader can now read webpage
contents, though navigation is still broken.