Bug 31920: Fix Security Level panel when its toolbar button moves to overflow
Simply removed the 'clever' caching of the underlying toolbar button
element. Now, each time the button is needed it is queried from the
underlying document.

A correct, but more complicated, fix would be to determine when the
toolbar button had been moved to the overflow menu and then update our
cached reference. However, this would be super brittle and would
break once another way to move the toolbar button is added in the
future.

In the end, caching something so infrequently queried is not worth the
complexity.